# backend/src/api-zod/generated/types/courseDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SkillPath AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CutoffHistoryPoint } from './cutoffHistoryPoint';

export interface CourseDetail {
  id: number;
  universityId: number;
  universityName: string;
  universityShortName?: string;
  degreeName: string;
  faculty: string;
  degreeType: string;
  durationYears: number;
  minimumZScore: number;
  stream: string;
  /** @nullable */
  description?: string | null;
  subjects: string[];
  skillsDeveloped: string[];
  /** @nullable */
  officialMinimumZScore?: number | null;
  /** @nullable */
  predictedMinimumZScore?: number | null;
  /** @nullable */
  officialAcademicYear?: string | null;
  /** @nullable */
  predictedAcademicYear?: string | null;
  /** @nullable */
  confidence?: string | null;
  /** @nullable */
  dataSource?: string | null;
  /** @nullable */
  district?: string | null;
  cutoffHistory?: CutoffHistoryPoint[];
}